+++ denominator-core/src/test/java/denominator/BaseRoundRobinLiveTest.java
@@ -48,25 +48,25 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         skipIfNoCredentials();
         Zone zone = skipIfNoMutableZone();
 
-        skipIfRRSetExists(zone, recordSet.getName(), recordSet.getType());
+        skipIfRRSetExists(zone, recordSet.name(), recordSet.type());
 
         rrsApi(zone).add(ResourceRecordSet.<Map<String, Object>> builder()
-                                              .name(recordSet.getName())
-                                              .type(recordSet.getType())
+                                              .name(recordSet.name())
+                                              .type(recordSet.type())
                                               .ttl(1800)
-                                              .add(recordSet.get(0)).build());
+                                              .add(recordSet.rdata().get(0)).build());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
         checkRRS(rrs.get());
-        assertEquals(rrs.get().getName(), recordSet.getName());
-        assertEquals(rrs.get().getTTL().get(), Integer.valueOf(1800));
-        assertEquals(rrs.get().getType(), recordSet.getType());
-        assertEquals(rrs.get().size(), 1);
-        assertEquals(rrs.get().get(0), recordSet.get(0));
+        assertEquals(rrs.get().name(), recordSet.name());
+        assertEquals(rrs.get().ttl().get(), Integer.valueOf(1800));
+        assertEquals(rrs.get().type(), recordSet.type());
+        assertEquals(rrs.get().rdata().size(), 1);
+        assertEquals(rrs.get().rdata().get(0), recordSet.rdata().get(0));
     }
 
     @Test(dependsOnMethods = "createNewRRS", dataProvider = "roundRobinRecords")
@@ -75,21 +75,21 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         Zone zone = skipIfNoMutableZone();
 
         rrsApi(zone).add(ResourceRecordSet.<Map<String, Object>> builder()
-                                              .name(recordSet.getName())
-                                              .type(recordSet.getType())
-                                              .add(recordSet.get(1)).build());
+                                              .name(recordSet.name())
+                                              .type(recordSet.type())
+                                              .add(recordSet.rdata().get(1)).build());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
         checkRRS(rrs.get());
-        assertEquals(rrs.get().getName(), recordSet.getName());
-        assertEquals(rrs.get().getType(), recordSet.getType());
-        assertEquals(rrs.get().size(), 2);
-        assertEquals(rrs.get().get(0), recordSet.get(0));
-        assertEquals(rrs.get().get(1), recordSet.get(1));
+        assertEquals(rrs.get().name(), recordSet.name());
+        assertEquals(rrs.get().type(), recordSet.type());
+        assertEquals(rrs.get().rdata().size(), 2);
+        assertEquals(rrs.get().rdata().get(0), recordSet.rdata().get(0));
+        assertEquals(rrs.get().rdata().get(1), recordSet.rdata().get(1));
     }
 
     @Test(dependsOnMethods = "addRecordToExistingRRS", dataProvider = "roundRobinRecords")
@@ -97,20 +97,20 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         skipIfNoCredentials();
         Zone zone = skipIfNoMutableZone();
 
-        rrsApi(zone).applyTTLToNameAndType(200000, recordSet.getName(), recordSet.getType());
+        rrsApi(zone).applyTTLToNameAndType(200000, recordSet.name(), recordSet.type());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
         checkRRS(rrs.get());
-        assertEquals(rrs.get().getName(), recordSet.getName());
-        assertEquals(rrs.get().getType(), recordSet.getType());
-        assertEquals(rrs.get().getTTL().get(), Integer.valueOf(200000));
-        assertEquals(rrs.get().size(), 2);
-        assertEquals(rrs.get().get(0), recordSet.get(0));
-        assertEquals(rrs.get().get(1), recordSet.get(1));
+        assertEquals(rrs.get().name(), recordSet.name());
+        assertEquals(rrs.get().type(), recordSet.type());
+        assertEquals(rrs.get().ttl().get(), Integer.valueOf(200000));
+        assertEquals(rrs.get().rdata().size(), 2);
+        assertEquals(rrs.get().rdata().get(0), recordSet.rdata().get(0));
+        assertEquals(rrs.get().rdata().get(1), recordSet.rdata().get(1));
     }
 
     @Test(dependsOnMethods = "applyTTL", dataProvider = "roundRobinRecords")
@@ -119,24 +119,24 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         Zone zone = skipIfNoMutableZone();
 
         rrsApi(zone).replace(ResourceRecordSet.<Map<String, Object>> builder()
-                                                  .name(recordSet.getName())
-                                                  .type(recordSet.getType())
+                                                  .name(recordSet.name())
+                                                  .type(recordSet.type())
                                                   .ttl(10000)
-                                                  .add(recordSet.get(0))
-                                                  .add(recordSet.get(2)).build());
+                                                  .add(recordSet.rdata().get(0))
+                                                  .add(recordSet.rdata().get(2)).build());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
         checkRRS(rrs.get());
-        assertEquals(rrs.get().getName(), recordSet.getName());
-        assertEquals(rrs.get().getType(), recordSet.getType());
-        assertEquals(rrs.get().getTTL().get(), Integer.valueOf(10000));
-        assertEquals(rrs.get().size(), 2);
-        assertEquals(rrs.get().get(0), recordSet.get(0));
-        assertEquals(rrs.get().get(1), recordSet.get(2));
+        assertEquals(rrs.get().name(), recordSet.name());
+        assertEquals(rrs.get().type(), recordSet.type());
+        assertEquals(rrs.get().ttl().get(), Integer.valueOf(10000));
+        assertEquals(rrs.get().rdata().size(), 2);
+        assertEquals(rrs.get().rdata().get(0), recordSet.rdata().get(0));
+        assertEquals(rrs.get().rdata().get(1), recordSet.rdata().get(2));
     }
 
     @Test(dependsOnMethods = "replaceExistingRRSUpdatingTTL", dataProvider = "roundRobinRecords")
@@ -145,20 +145,20 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         Zone zone = skipIfNoMutableZone();
 
         rrsApi(zone).remove(ResourceRecordSet.<Map<String, Object>> builder()
-                                                 .name(recordSet.getName())
-                                                 .type(recordSet.getType())
-                                                 .add(recordSet.get(2)).build());
+                                                 .name(recordSet.name())
+                                                 .type(recordSet.type())
+                                                 .add(recordSet.rdata().get(2)).build());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
         checkRRS(rrs.get());
-        assertEquals(rrs.get().getName(), recordSet.getName());
-        assertEquals(rrs.get().getType(), recordSet.getType());
-        assertEquals(rrs.get().size(), 1);
-        assertEquals(rrs.get().get(0), recordSet.get(0));
+        assertEquals(rrs.get().name(), recordSet.name());
+        assertEquals(rrs.get().type(), recordSet.type());
+        assertEquals(rrs.get().rdata().size(), 1);
+        assertEquals(rrs.get().rdata().get(0), recordSet.rdata().get(0));
     }
 
     @Test(dependsOnMethods = "removeRecordFromExistingRRS", dataProvider = "roundRobinRecords")
@@ -167,41 +167,41 @@ public abstract class BaseRoundRobinLiveTest extends BaseProviderLiveTest {
         Zone zone = skipIfNoMutableZone();
 
         rrsApi(zone).remove(ResourceRecordSet.<Map<String, Object>> builder()
-                                                 .name(recordSet.getName())
-                                                 .type(recordSet.getType())
-                                                 .add(recordSet.get(0)).build());
+                                                 .name(recordSet.name())
+                                                 .type(recordSet.type())
+                                                 .add(recordSet.rdata().get(0)).build());
 
         Optional<ResourceRecordSet<?>> rrs = rrsApi(zone)
-                .getByNameAndType(recordSet.getName(), recordSet.getType());
+                .getByNameAndType(recordSet.name(), recordSet.type());
 
         assertFalse(rrs.isPresent(),
-                format("recordset(%s, %s) still present in %s", recordSet.getName(), recordSet.getType(), zone));
+                format("recordset(%s, %s) still present in %s", recordSet.name(), recordSet.type(), zone));
     }
 
     @Test(dataProvider = "roundRobinRecords")
     private void deleteRRS(ResourceRecordSet<?> recordSet) {
         skipIfNoCredentials();
         Zone zone = skipIfNoMutableZone();
-        String recordName = recordSet.getName().replace("." + zone.name(), "-delete." + zone.name());
+        String recordName = recordSet.name().replace("." + zone.name(), "-delete." + zone.name());
 
-        skipIfRRSetExists(zone, recordName, recordSet.getType());
+        skipIfRRSetExists(zone, recordName, recordSet.type());
 
         rrsApi(zone).add(
-                ResourceRecordSet.<Map<String, Object>> builder().name(recordName).type(recordSet.getType())
+                ResourceRecordSet.<Map<String, Object>> builder().name(recordName).type(recordSet.type())
                         .addAll(recordSet).build());
 
-        Optional<ResourceRecordSet<?>> rrs = rrsApi(zone).getByNameAndType(recordName, recordSet.getType());
+        Optional<ResourceRecordSet<?>> rrs = rrsApi(zone).getByNameAndType(recordName, recordSet.type());
 
-        assertPresent(rrs, zone, recordSet.getName(), recordSet.getType());
+        assertPresent(rrs, zone, recordSet.name(), recordSet.type());
 
-        rrsApi(zone).deleteByNameAndType(recordName, recordSet.getType());
+        rrsApi(zone).deleteByNameAndType(recordName, recordSet.type());
 
-        String failureMessage = format("recordset(%s, %s) still exists in %s", recordName, recordSet.getType(), zone);
-        assertFalse(rrsApi(zone).getByNameAndType(recordName, recordSet.getType()).isPresent(), failureMessage);
-        assertFalse(any(rrsApi(zone).iterator(), and(nameEqualTo(recordName), typeEqualTo(recordSet.getType()))),
+        String failureMessage = format("recordset(%s, %s) still exists in %s", recordName, recordSet.type(), zone);
+        assertFalse(rrsApi(zone).getByNameAndType(recordName, recordSet.type()).isPresent(), failureMessage);
+        assertFalse(any(rrsApi(zone).iterator(), and(nameEqualTo(recordName), typeEqualTo(recordSet.type()))),
                 failureMessage);
 
         // test no exception if re-applied
-        rrsApi(zone).deleteByNameAndType(recordName, recordSet.getType());
+        rrsApi(zone).deleteByNameAndType(recordName, recordSet.type());
     }
 }